refactor: extract command execution code into CommandHandler.kt
